+++ ErsatzTV.Application/ProgramSchedules/ProgramScheduleItemOneViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using ErsatzTV.Application.Filler;
+using ErsatzTV.Application.Graphics;
 using ErsatzTV.Application.MediaCollections;
 using ErsatzTV.Application.MediaItems;
 using ErsatzTV.Application.Watermarks;
@@ -31,6 +32,7 @@ public record ProgramScheduleItemOneViewModel : ProgramScheduleItemViewModel
         FillerPresetViewModel tailFiller,
         FillerPresetViewModel fallbackFiller,
         List<WatermarkViewModel> watermarks,
+        List<GraphicsElementViewModel> graphicsElements,
         string preferredAudioLanguageCode,
         string preferredAudioTitle,
         string preferredSubtitleLanguageCode,
@@ -57,6 +59,7 @@ public record ProgramScheduleItemOneViewModel : ProgramScheduleItemViewModel
         tailFiller,
         fallbackFiller,
         watermarks,
+        graphicsElements,
         preferredAudioLanguageCode,
         preferredAudioTitle,
         preferredSubtitleLanguageCode,